requirements.txt
openai
requests
python-dotenv
customtkinter